Restructured gql folder and refactored automatic imports
