--- app/graphql/healthCheck/queries.js
@@ -1,12 +1,0 @@
-const { gql } = require('apollo-server');
-
-module.exports = {
-  queries: {
-    healthCheck: () => process.uptime()
-  },
-  schema: gql`
-    extend type Query {
-      healthCheck: String!
-    }
-  `
-};
