+++ app/graphql/healthCheck/resolvers.js
@@ -1,0 +1,5 @@
+module.exports = {
+  Query: {
+    healthCheck: () => process.uptime()
+  }
+};
