+++ app/graphql/healthCheck/schemas.js
@@ -1,0 +1,9 @@
+const { gql } = require('apollo-server');
+
+const rootTpyes = gql`
+  extend type Query {
+    healthCheck: String!
+  }
+`;
+
+exports.schemas = [rootTpyes];
