+++ app/graphql/scripts.js
@@ -1,83 +1,29 @@
 const fs = require('fs'),
-  path = require('path');
-
-const INDEX_FILE = 'index.js',
-  SCRIPTS_FILE = 'scripts.js';
-
-const generateIndex = dirent =>
-  fs.readdirSync(path.join(__dirname, dirent.name)).reduce(
-    (singleModule, file) => {
-      const fileName = file.replace(/.js/gi, '');
-      const definitions = require(path.join(__dirname, dirent.name, fileName));
-      singleModule[fileName] = definitions[fileName] ? definitions[fileName] : definitions;
-      if (definitions.schema) {
-        singleModule.schemas.push(definitions.schema);
-      }
-      return singleModule;
-    },
-    { schemas: [] }
-  );
+  path = require('path'),
+  merge = require('lodash.merge');
 
 exports.importEverything = () =>
   fs.readdirSync(__dirname, { withFileTypes: true }).reduce(
     (imports, dirent) => {
       if (dirent.isDirectory()) {
-        const index = generateIndex(dirent);
-        imports.modules.push(index);
-        return imports;
-      } else if (dirent.name === INDEX_FILE || dirent.name === SCRIPTS_FILE) {
+        fs.readdirSync(path.join(__dirname, dirent.name)).forEach(file => {
+          const fileName = file.replace(/.js/gi, '');
+          if (fileName === 'schemas') {
+            const { schemas } = require(path.join(__dirname, dirent.name, fileName));
+            imports.schemas.push(...schemas);
+          }
+          if (fileName === 'resolvers') {
+            const resolvers = require(path.join(__dirname, dirent.name, fileName));
+            merge(imports.resolvers, resolvers);
+          }
+          if (fileName === 'middlewares') {
+            const middlewares = require(path.join(__dirname, dirent.name, fileName));
+            merge(imports.middlewares, middlewares);
+          }
+        });
         return imports;
       }
-      imports[dirent.name.replace(/.js/gi, '')] = require(path.join(__dirname, dirent.name));
       return imports;
     },
-    { modules: [] }
+    { schemas: [], resolvers: {}, middlewares: {} }
   );
-
-exports.destructureModules = modules => {
-  const rootObject = { Query: {}, Mutation: {}, Subscription: {} };
-  return modules.reduce(
-    (destructuredModules, currentModule) => {
-      destructuredModules.schemas.push(...currentModule.schemas);
-
-      const { fieldResolvers } = currentModule.resolvers ? currentModule.resolvers : {};
-      destructuredModules.resolvers = {
-        ...destructuredModules.resolvers,
-        Query: {
-          ...destructuredModules.resolvers.Query,
-          ...currentModule.queries
-        },
-        Mutation: {
-          ...destructuredModules.resolvers.Mutation,
-          ...currentModule.mutations
-        },
-        Subscription: {
-          ...destructuredModules.resolvers.Subscription,
-          ...currentModule.subscriptions
-        },
-        ...fieldResolvers
-      };
-
-      if (currentModule.middlewares) {
-        destructuredModules.middlewares = {
-          ...destructuredModules.middlewares,
-          Query: {
-            ...destructuredModules.middlewares.Query,
-            ...currentModule.middlewares.queries
-          },
-          Mutation: {
-            ...destructuredModules.middlewares.Mutation,
-            ...currentModule.middlewares.mutations
-          },
-          Subscription: {
-            ...destructuredModules.middlewares.Subscription,
-            ...currentModule.middlewares.subscriptions
-          },
-          ...currentModule.middlewares.fieldMiddlewares
-        };
-      }
-      return destructuredModules;
-    },
-    { schemas: [], resolvers: rootObject, middlewares: rootObject }
-  );
-};
