+++ app/graphql/users/middlewares.js
@@ -8,8 +8,8 @@ const createUser = (resolve, root, args) => {
 
 module.exports = {
   // Here you add all the middlewares for the mutations, queries or field resolvers if you have any
-  mutations: {
+  Mutation: {
     createUser
   },
-  fieldMiddlewares: { User: {} }
+  User: {}
 };
