--- app/graphql/users/subscriptions.js
@@ -1,15 +1,0 @@
-const { gql } = require('apollo-server'),
-  { userLoggedIn } = require('../events');
-
-module.exports = {
-  subscriptions: {
-    onLogin: {
-      subscribe: userLoggedIn.iter
-    }
-  },
-  schema: gql`
-    extend type Subscription {
-      onLogin: String!
-    }
-  `
-};
