+++ package.json
@@ -63,6 +63,7 @@
     "graphql-middleware": "^3.0.2",
     "graphql-tools": "^4.0.4",
     "jwt-simple": "^0.5.1",
+    "lodash.merge": "^4.6.2",
     "pg": "^7.4.1",
     "rollbar": "^2.3.9",
     "sequelize": "^4.44.0",
